requirements.txt
# --- Core API ---
fastapi==0.115.0
uvicorn[standard]==0.30.5

# --- ORM / DB ---
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9

# --- Validation & config ---
pydantic==2.9.2
python-dotenv==1.0.1

# --- Messaging (RabbitMQ, async) ---
aio-pika==9.4.1

# --- Monitoring / Metrics ---
prometheus-client==0.20.0

# --- Tests ---
pytest==8.3.2
pytest-cov==5.0.0
httpx==0.27.2

# --- Auth ---
pyjwt[crypto]
